+++ dwds/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 19.0.1-dev
 
+- Do not show async frame errors on evaluation. - [#2073](https://github.com/dart-lang/webdev/pull/2073)
 - Refactor code for presenting record instances. - [#2074](https://github.com/dart-lang/webdev/pull/2074)
 
 ## 19.0.0